move metadata calculation to datasource storage
