+++ app/server/appsmith-server/src/main/java/com/appsmith/server/datasourcestorages/base/DatasourceStorageServiceCECompatibleImpl.java
@@ -4,6 +4,8 @@ import com.appsmith.server.helpers.PluginExecutorHelper;
 import com.appsmith.server.plugins.base.PluginService;
 import com.appsmith.server.repositories.DatasourceStorageRepository;
 import com.appsmith.server.services.AnalyticsService;
+import com.appsmith.server.services.ConfigService;
+import com.appsmith.server.services.OrganizationService;
 import com.appsmith.server.solutions.DatasourcePermission;
 import org.springframework.stereotype.Service;
 
@@ -16,7 +18,16 @@ public class DatasourceStorageServiceCECompatibleImpl extends DatasourceStorageS
             DatasourcePermission datasourcePermission,
             PluginService pluginService,
             PluginExecutorHelper pluginExecutorHelper,
-            AnalyticsService analyticsService) {
-        super(repository, datasourcePermission, pluginService, pluginExecutorHelper, analyticsService);
+            AnalyticsService analyticsService,
+            ConfigService configService,
+            OrganizationService organizationService) {
+        super(
+                repository,
+                datasourcePermission,
+                pluginService,
+                pluginExecutorHelper,
+                analyticsService,
+                configService,
+                organizationService);
     }
 }
